add tests home page genadrop GD_HP012, GD_HP025, GD_HP038
